+++ PKHeX.Core/MysteryGifts/WC6.cs
@@ -16,9 +16,11 @@ namespace PKHeX.Core
             Data = (byte[])(data?.Clone() ?? new byte[Size]);
             if (Data.Length == SizeFull)
             {
-                if (Data[0x205] == 0)
-                    Data = new byte[Data.Length]; // Invalidate
-                Data = Data.Skip(SizeFull - Size).ToArray();
+                byte[] wc6 = new byte[Size];
+                if (Data[0x205] != 0) // Valid data
+                    Array.Copy(Data, SizeFull - Size, wc6, 0, wc6.Length);
+                Data = wc6;
+                
                 DateTime now = DateTime.Now;
                 Year = (uint)now.Year;
                 Month = (uint)now.Month;
